Treat test warnings as errors
I don't imagine this is going to make me any friends, but the aim is to
force us to be more proactive about fixing deprecation warnings, and
ensuring that we're generally updated old code and not letting it
stagnate.

At this point in time, I was able to run the test suite with Numpy 1.19
and SciPy 1.6 on Mac without any warnings propagating through.