Build docs on PR using GitHub Actions
This is intended as a CI test of the documentation build.  It treats
warnings in the Sphinx build as errors, so we don't accidentally break
anything on new PRs.

This is not perfect.  Most errors will come from problems in the main
QuTiP repository's API documentation (docstrings).  These will only be
caught on the next documentation build, which may cause some minor
problems for contributors to the docs repository, but hopefully we will
at least catch them and get them fixed again.

Once qutip-doc is merged into qutip/qutip these strange cross-repository
annoyances will go away.