Exclude documentation repository from CodeClimate
The exclusion of the "dist" and "tests" directories is not new, despite
appearances.  Before this commit, we did not specify any exclude
patterns, so CodeClimate used its defaults, which includes those two
folders.

This commit isn't intending to actually change the behaviour, so it
simply codifies the defaults for now.  They can be changed later if
desired.