Change 'release' flag to be derived from VERSION
Previously setup.py decided if it was in "release mode" by looking at
command-line flags or environment variables.  This meant that
distributed sdists could not tell that they were supposed to be building
in release mode, and so would add on additional local version
information.

Instead, we change our release procedures such that the VERSION file has
a '.dev' postfix if this is a development release, and it doesn't have
one if it isn't.  Development branches such as 'master' and 'dev.major'
will always have this development postfix, but release branches won't.