Rewrite entire git procedure for release
In theory nothing has changed in the actual procedures compared to the
previous version, but this rewrites most of the guide to make things
much clearer and easier to understand.  This was made following the
release of 4.6.1.

There are still some changes to be made to the methods, but this
involves additional code changes that aren't in place yet.  The
document as-is describes the current release procedures.

Co-authored-by: Simon Cross <hodgestar@gmail.com>