Move tests to GitHub Actions
This moves our entire testing apparatus over to GitHub Actions, since
Travis has changed its pricing model and we can no longer rely on it.
This actually has several advantages for us right now:
 - better integration with other GitHub Actions workflows
 - more concurrent jobs
 - more access to Windows and macOS machines
 - easier job dependency graphs within our resources

This first commit does not add in any Windows tests because they're
currently known to fail, but the intention is to turn them on.  The
blocking issue is currently a failure of the multiprocessing
capabilities, which also affects mcsolve.

We can now reasonably run coverage reporting on every test run, and
upload all of them to Coveralls at the end, so we get accurate stats on
what lines were actually tested.