BoFiN: Modularization and fermionic test
Started on some slight modularization of both process_input() functions.

Also added a basic fermionic solver test based against the the single-impurity current solution (which was in example 4b).  This is kind of clunky because it needs lots of utility functions from that example.  I should just bite the bullet and include these in the heom.py