Rewrite Husimi-Q calculation and tests
This is a complete rewrite of the Husimi Q calculations, to make a more
formal split between the iterative (low-memory) and precomputed (faster)
algorithms, giving full access to both of them.  This is built on the
shoulders of 65d0ec08, but is an entire reimplementation, giving it
quite a bit of a speed-up as well.

The concept of the precomputed matrix is encapsulated into a class with
a cached tensor, so that the same precomputed system can be used for
many states of different sizes, and autoexpands (up to the memory limit)
to accommodate larger states.

It also adds rather more error checking on its inputs, and adds a
complete testing suite for the new implementation.

Timings
-------

These benchmarks were taken on a Macbook Pro 2.9GHz i5 (2015) with 8GB
of RAM.  The different tests had different numbers of phase-space
coordinates (the first number in the first column) in each dimension,
and the number after 'ket' or 'dm' is the Hilbert space dimension.

The columns are:
    original:
        The state of the code before this and the previous commit
    prev def:
        Direct calls of `qutip.qfunc` with the given state and
        phase-space coordinates, but everything else at the default for
        the previous commit (65d0ec08)
    prev pre:
        Calls of `qutip.qfunc(..., precompute=pre)`, where the
        calculation of `pre` was done outside the timing.  This times
        how long it would take to call in a loop many times.
    cur def:
        Direct calls of `qutip.qfunc` with the given state and
        phase-space coordinates, but everything else at the default for
        this commit.
    cur pre:
        Calls to a pre-constructed `qutip.QFunc` instance (which was
        excluded from the timings) for this commit.  This times how long
        it would take to call in a loop many times.

            │ original  │ prev def  │ prev pre  │  cur def  │  cur pre
────────────┼───────────┼───────────┼───────────┼───────────┼──────────
11,  ket 5  │ 126.07 µs │ 133.02 µs │  89.63 µs │ 108.95 µs │  25.12 µs
11,  ket 32 │ 192.49 µs │ 197.88 µs │  90.91 µs │ 172.13 µs │  29.33 µs
11,  dm 5   │ 481.68 µs │ 440.45 µs │ 316.93 µs │ 304.86 µs │ 161.00 µs
11,  dm 32  │   4.69 ms │   1.87 ms │   1.66 ms │   1.07 ms │ 823.49 µs
            │           │           │           │           │
101, ket 5  │ 570.59 µs │ 566.65 µs │ 501.86 µs │ 463.96 µs │ 329.25 µs
101, ket 32 │   1.61 ms │   1.43 ms │ 751.97 µs │   1.34 ms │ 507.20 µs
101, dm 5   │   2.51 ms │   3.51 ms │   2.20 ms │   2.62 ms │   1.72 ms
101, dm 32  │  42.91 ms │  26.84 ms │  19.73 ms │  23.30 ms │  16.30 ms
            │           │           │           │           │
401, ket 5  │  10.56 ms │  13.50 ms │   7.91 ms │  11.95 ms │   5.28 ms
401, ket 32 │  34.44 ms │  43.73 ms │  10.37 ms │  41.17 ms │   7.70 ms
401, dm 5   │  46.85 ms │  57.51 ms │  33.19 ms │  52.30 ms │  26.71 ms
401, dm 32  │   1.05  s │ 411.32 ms │ 273.93 ms │ 428.32 ms │ 246.59 ms