Clarify attribution line of licence
This change is not intended to remove Paul and Robert as copyright
holders, nor to remove our acknowledgement of their original
contributions.  They are still copyright holders, and remain QuTiP
contributors.

Instead, this change is meant to more accurately represent the copyright
holders to readers of the licence.  Previously, and currently, QuTiP
does not have a contributor licensing agreement, so all contributors
still own the copyright to their contributions, regardless of what was
stated in the licence file.

The years of the licence are also made explicit.  There have been public
releases of QuTiP, with modifications, in each of the years from 2011 to
2021 inclusive, and so copyright is asserted explicitly for each of
these.  It is still the intention to update the years of the copyright
as new modifications and releases are made.

This patch is not intended to change the copyright holders, only to
clarify who they are and who they have always been.

See qutip/qutip#1579