Add licensing information to documentation builds
This clarifies that the documentation _text_ is licensed under the
Creative Commons Attribution 3.0 Unported licence, while code samples
and the reproductions of QuTiP source code are licensed under the terms
of the 3-clause BSD licence.

This includes reproduction of the terms of both licenses, and the same
update to the copyright holder attribution.  The addition of the
`_copyright` rst target means that every documentation page gains a
hyperlink to the copyright notices in both the sidebar and the footer.