Remove licence information from individual files
Having the licence information copied into each individual file is
redundant, and while it might make things marginally easier for people
copying entire files out of our codebase, it is more difficult for us to
manage and to update the attributions and years.  It is not a
requirement of the BSD 3-clause licence to include the text in each
individual file, and other similar projects using the same licence do
not do this.  For example:
    - numpy
    - scipy
    - pandas
    - scikit-learn
all have the same licence, and none of them include the text in each
file.

Removing the additional information should also make it easier to see
places in the code which are licensed under different terms.  Note that
there are currently no parts of the QuTiP codebase that are licensed
under different terms.

This not intended to change any licence terms of the package,
documentation or associated data.

See qutip/qutip#1579 for more information and discussion.