faster state_number_enumerate
When the number of excitations is limited, instead of iterating over all
states and discarding the ones with too many excitations, directly
choose the limits to only iterate over allowed states. As a small
additional optimization, do not redo the same sum every time, but keep
track of the sum within the algorithm.