Minor doc fix in QubitCircuit
The `num_cbits` wasn't documented in the QubitCircuit docstring. This adds it.

Took me a second to figure out what it was (as a novice in quantum computing), so I figured I'd save some other novice that second.

FWIW: `reverse_states` isn't documented either, but I'm not confident I can write an accurate description of what that is, otherwise I would have put it here too.