Added Tests to __init__ module to allow importing differently.
We can now import the tests using:
```
import qutip.tests.core.data as testing
class TestAdd_2(testing.Add):
	.....
```

Allowing this import method means that the "TestAdd" that will be run is
the "TestAdd_2". Other options is to import TestAdd directly and delete
it at the end of the file. I found this solution more elegant.