Handle change in configuration attributes for numpy 1.22.0
In the recently released numpy 1.22.0 the __config__ attribute for the
published wheels (at least on the linux platforms I'm using) have
removed the 'blas_opt_info' attribute and are instead publishing
identical information in the 'blas_ilp64_opt_info' attribute. Running
qutip with the latest numpy release on these platforms was causing an
attribute error when _blas_info() is called as
numpy.__config__.blas_opt_info doesn't exist in the installed numpy.
This commit fixes this by first checking if the new name exists and
using that to get the blas opt info, and then trying the name available
in previous releases. This should maintain compatibility with older
numpy releases but fix the issue when running with 1.22.0.