FlimeSolve steady state solver
Updated Flimesolve with a steady state solver. This uses the rate matrix as an analogue for the Monodromy matrix to find the steady state of the system in the full secular approximation. Not sure if adding in more time-dependence (if possible) would do anything, as I think that would just cause the solution to orbit/rotate around the time independent result.