Adding Back in the FFT
The FFT is the only way to bring FLiMESolve to agree with MESolve. Additionally, FMMESolve displays the same issue as FLiMESolve when the driving amplitude is brought to zero. This makes me think that there's an eigenvalue/vector ordering issue here. Probably degeneracies....somewhere? It's also the only way to get all the indices to have signs that agree with the theoretical FLiMESolve.